Implement removeElement function using erase method
Implement function to remove specified element from vector.
Uses the vector erase() method to remove elements matching the target value by maintaining two pointers i and j for current position and elements examined counter respectively. 
When a matching element is found , it's erase and the array(nums) automatically shifts and when a non-matching element is found, both pointers get an increment. 
Then return the final size of the array.