Create 3. Longest Substring Without Repeating Characters.cpp
This solution calculates the length of the longest substring without repeating characters by utilizing a sliding window(i->start, j ), a dynamic vector to keep track of current characters, and a 256 size boolean lookup table for fast duplicate detection. The window is expanded when all the characters are unique; otherwise, it resets. Time complexity is efficient for typical inputs (like size of string is 0 or 1).